Add support for `custom` FFI attribute.
Some types (like `wise_units::Unit`) are a bit different for FFI; they
have an internal structure that could be exposed, but it's not really a
useful interface (e.g., we care about a `wise_units::Unit` as a thing
that can be initialized from a UCUM expression, and we need to be able
to read a UCUM expression out of it -- we don't care about its internal
`terms`, or the fields of the `Term` type, etc.).

So, we need a way to let a type be `ffi_derive` (so it can take
advantage of all the boilerplate stuff + get a consumer generated for
itself), but provide its own implementation of an initializer and getter
functions. This commit adds support for this, but it's *very* brittle/
holey. Needs a ton of work in DEV-13175 to do it properly.

Add cargo config file so that this crate knows where to find its deps
without needing a special env configuration.

Also, make macros a little more hygienic so that we don't need as many
weird imports.