Redesign FFI optionals to avoid nested struct layout issues.
Ran into some memory layout issues on `aarch64-apple-ios`, where Swift
would copy FFI function arguments into different positions than where
Rust expected them. This led to some exciting undefined behavior where
we might try to start reading a `String` in the middle of the padding
between argument values, or interpret some series of `c_char` bytes as
an impossibly large vec count (which led to malloc complaining about not
being able to alloc 7 GBs on an iPad).

This needs more investigation (since either Swift isn't sound when making
calls to C APIs in some cases or Rust has an error in its alignment for
the `aarch64-apple-ios` target), but for now, we're making the FFI safe
for devices by putting optional types behind pointers.

So, if we have an `Option<u16>`, it'll actually be exposed as (and
received as when used as an initializer field) a `*const u16`, which the
consumer can then inspect for null-ness and dereference if a value is
found.

Refactor the `TimeStamp` type as well so that it's passed by reference
(same reason; we don't want to deal with embedding any structs in
function parameters right now).

Make the ownership rules more consistent. Some types were still being
cloned as part of initializing them in Rust (in cases where the consumer
typically retains a pointer in an object that may outlive the
initializer call). Now all data passed to a Rust FFI initializer is
considered `moved`; Rust will retake ownership of the memory of
arguments. To support this, some complex types from the consumer side
are now cloned (by calling `clone()` on the instance to which they
retain a pointer) when calling `toRust()`, so that the consumer type can
continue its normal lifespan while Rust takes ownership of the instance
it's passed. (Strings are the one exception to this because the language
already has support for initializing them from a view of foreign data, so
there's no benefit to init'ing them in Rust separately.)