DEV-13316 Improve alias resolution
Add an attribute macro for getting alias resolution details out of a
module early on (during proc macro expansion, before type analysis).
Write it to a file in the out dir, and read it when deriving FFI.

Alternatives tried:
1. Using the attribute macro to generate an alias_resolution macro
(no good because that macro isn't callable during the proc macro step
where derive is invoked)
2. Using a global variable instead of writing a file (no good because
the macro runs as a separate process for each crate, so alias
resolution info acquired while parsing agrian_types is not available
when parsing commodity_locationizer).