DEV-15924 Support separating client frameworks by crate.
Drop the rest of the FFI snake-casing. If we can't do it consistently
for the generated FFI code and the generated consumer code, we shouldn't
bother doing it at all. :( Might revisit this eventually; we'd need to
write something with consistent rules that can be applied to proc macros
and standard code.

Make the Swift consumer code public as needed. On the consumer side,
we're setting up one framework per FFI crate, so the Swift interface
needs to be public so other frameworks and the application can see it.

Add support for importing a common consumer framework in generated
consumer code. Fundamental generated code (like the FFI glue for numeric
and string types, etc) will generally go in a single common consumer
framework, which all other consumer frameworks will need to import. This
lets the consumer tell us what the name of that framework is through an
environment variable so that we can include the appropriate import
statement in the generated consumer code.

Add support for defining additional imports that may be required in the
generated consumer code. Now that the generated code is not all going in
a single monolithic framework, certain types/functions may need to
define the types that they require so that we can import them.