DEV-16437 Harden alias resolution
Remove most of the panics from `ffi_internals::alias_resolution` and
bubble up whatever goes wrong into a new `Error` type. We're still
unwrapping _that_ error, but now it's unwrapped in the attribute macro
itself, so it's a little easier to follow and will be easier to do
better error handling later.

Change `resolve_type_alias` to take an optional path to the resolution
file. This lets us keep and reuse a single lock on the `ALIAS_MAP_PATH`
mutex, instead of manually dropping it (unlocking the mutex) before each
recursive call.

Replace the use of `serde_json::to_writer` with `std::fs::write`. The
serde_json method doesn't necessarily replace the entire contents of the
file, which I think led to the broken JSON (if a larger `HashMap` was
already written, writing a smaller one would overwrite _part_ of that,
leaving a weirdly incomplete structure).

Clean up some of the file read/write calls and the `HashMap` mutations.
Fix some warnings.