DEV-15926 agrian_units support
Add an attribute macro for exposing non-impl functions.

Add support for borrowed function parameters. (Kinda hacky at the
moment, need to do a lot of cleanup in `native_type_data` and consider
refactoring a lot of that code.)

Very work-in-progress, but it supports deriving an FFI for the ops
functions on a `Measurement` now. This breaks the commodity_locationizer
implementation; that's probably a result of the borrow implementation in
the macro being very half-baked.