DEV-17287 Fix generated whitespace.
Use width formatter in a few places and clean up the structure of the
generated string.

There's a whole bunch more we could do with the width formatter, but I'm
going to hold off on that until we get around to building a quote-like
representation for Swift, since the logic of increasing and decreasing
indentation levels really belongs to something like that, not our
hardcoded class and function definitions.