Change the names of the generated FFI init/free/clone methods.
When we upload an app binary for release, Apple does some symbol use
sniffing to check whether we're calling any of their private APIs. We're
not, but one generated symbol (`field_init`) matched one of their
private API symbols, so they flagged us for calling it. So, we're just
changing the names of those methods here to avoid confusing their tool.

(As an aside, this serves as kind of a nice proof-of-concept for making
changes in one place that update the FFI and the consumer code. The only
thing we have to change manually in app code is a hardcoded reference to
a few of the FFI functions for really specialized tests.)