Clean up github actions.
Switch to cargo-cache.

Nuke the unnecessary SSH key.

Don't try to test the proc macros on an iOS simulator; that's silly, we
don't care whether you can execute proc macro code on the simulator.
(And it breaks with `trybuild` because that needs to write to disk, so
we'd have to use `cargo-dinghy` to run these. Which is doable, but
again, silly because this code isn't meant to execute on iOS, it's meant
to run when building an iOS-compatible library.)