Fix for syn removing `__TestExhaustive`.
Most recent version of `syn` removed the `__TestExhaustive` variant in
favor of a new lint, which is breaking downstream builds if their
`Cargo.lock` happens to pick up `syn` 1.0.90. That lint still isn't
stable, so we won't know if syn introduces a new variant there, but the
fallback is fine (the impl item for whatever new language feature would
just be ignored).

(There are some lingering warnings and it'd be nice to bring these
crates up to the 2021 edition, but I don't have time at the moment.)