DEV-34559 Allow missing FFI_COMMON_FRAMEWORK env var
We use this env var to prepend `import SomeFramework` to generated,
non-core consumer files. This is useful for a client that wants to split
the crates they consume into multiple native frameworks, but a) that's
not needed for mobile now that things are getting trimmed, and b) we
don't need to be opinionated about how consumers organize the generated
code.